requirements.txt
libsass
django-sass-processor
django-crispy-forms
bbcode
markdown
Pygments
bcrypt
